Signal connection manager
Basic first version. Does not work at all yet, but the idea should
be pretty straight forward: It handles the control messaging and
maintains a state for each "control connection". It spins data
connections when needed and uses packet manager to build packets
without knowledge of the bit level stuff.